Do not invoke unix_chkpwd helper in pam_unix.so
A now about year-old change [^1] has broken NGINX PAM auth
when NGINX is run as a (somewhat) hardened systemd service
almost completely; unless the service is assigned CAP_SETUID
(read: sudo) rights, it will always fail to authenticate a client
because it will *always* try to launch the helper binary as root,
even if it has read access to /etc/shadow.

With this patch, the checking logic directly in `pam_unix.so` is
also compiled with `-DHELPER_COMPILE`, meaning that the .so will
always try to verify the password itself instead of relying on
the helper binary, removing the need for `setuid(0)`.

[^1]: https://github.com/linux-pam/linux-pam/commit/b3020da7da384d769f27a8713257fbe1001878be#diff-d7ff6757d4f74b58323438695015c2f71e3d30c90033a14daae54e4b38faacafR254